exception handling extended

+++ src/main/java/com/jolyvert/deliveryapp/service/CustomerService.java
@@ -2,6 +2,9 @@ package com.jolyvert.deliveryapp.service;
 
 
 import com.jolyvert.deliveryapp.dto.RegisterCustomerDto;
+import com.jolyvert.deliveryapp.exception.CustomerException;
+import com.jolyvert.deliveryapp.exception.FoodCartException;
+import com.jolyvert.deliveryapp.exception.LoginException;
 import com.jolyvert.deliveryapp.model.Customer;
 import com.jolyvert.deliveryapp.repository.CustomerRepository;
 import com.jolyvert.deliveryapp.repository.FoodCartRepository;
@@ -20,10 +23,15 @@ public class CustomerService {
         this.foodCartService = foodCartService;
     }
 
-    public Customer createCustomer(RegisterCustomerDto registerCustomerDto) {
+    public Customer createCustomer(RegisterCustomerDto registerCustomerDto) throws CustomerException {
+
+        if(customerRepository.existsByEmail(registerCustomerDto.getEmail())) {
+            throw new CustomerException("Customer already exists");
+        }
 
         Customer customer = new Customer();
         customer.setName(registerCustomerDto.getName());
+        customer.setTelephoneNumber(registerCustomerDto.getTelephoneNumber());
         customer.setEmail(registerCustomerDto.getEmail());
         customer.setPassword(registerCustomerDto.getPassword());
         customer.setAddress(registerCustomerDto.getAddress());
@@ -38,7 +46,12 @@ public class CustomerService {
         return customerRepository.findAll();
     }
 
-    public String deleteCustomer(Long id) {
+    public String deleteCustomer(Long id) throws CustomerException {
+
+        if(!customerRepository.existsById(id)) {
+            throw new CustomerException("Customer does not exist");
+        }
+
         customerRepository.delete(customerRepository.findById(id).orElseThrow(() -> new RuntimeException("Customer not found")));
         return "Customer deleted";
